Updates on the url-shortener app

+++ url-shortener-backend/src/services/url.service.ts
@@ -85,7 +85,7 @@ export class UrlService {
       SELECT 
         u.short_code,
         u.original_url,
-        COUNT(c.id)::INTEGER AS clicks
+        COUNT(c.id) AS clicks
       FROM urls u
       LEFT JOIN clicks c ON c.url_id = u.id
       WHERE u.user_id = $1
